#42 upgrade to 6.0 on the way to gradle 8.x
- fix deprecations
The archiveName property has been deprecated.
This is scheduled to be removed in Gradle 7.0.
Please use the archiveFileName property instead.
1 usage
The baseName property has been deprecated.
This is scheduled to be removed in Gradle 7.0.
Please use the archiveBaseName property instead.
1 usage
The destinationDir property has been deprecated.
This is scheduled to be removed in Gradle 7.0.
Please use the destinationDirectory property instead.
1 usage