Fix tasks for multipage formats
Task creation for the individual files of each format was previously
handled during the execution phase instead of the configuration phase,
leading for the tasks to be ignored.

Furthermore, the build had to be split up into multiple gradle runs
as the result of the generateDocbookMP task is not yet available
during the configuration phase, thus no subsequent tasks were generated.